Fix workflow worker import boundaries
