+++ apps/worker/src/workflows/workspace-gate.ts
@@ -1,4 +1,3 @@
-import { createHash } from "node:crypto";
 import type { PrePrCheckConfig } from "../pre-pr-checks/config.js";
 import {
   parseVerifiedWorkspaceManifest,
@@ -56,30 +55,6 @@ export class WorkspaceGateError extends Error {
 
 export function invalidateWorkspaceGate(state: WorkspaceGateState): void {
   state.prePrGate = null;
-}
-
-/**
- * Canonical fingerprint input includes the complete trusted manifest and one
- * HEAD per repository in manifest order. Clean-worktree verification happens
- * before this pure helper is called.
- */
-export function fingerprintWorkspaceState(
-  workspaceManifest: WorkspaceManifest,
-  headShas: readonly string[],
-): string {
-  if (headShas.length !== workspaceManifest.repositories.length) {
-    throw new Error("Workspace fingerprint requires one HEAD for every repository");
-  }
-  const payload = {
-    version: 1,
-    workspaceManifest,
-    repositories: workspaceManifest.repositories.map((repo, index) => ({
-      provider: repo.provider,
-      repoPath: repo.repoPath,
-      headSha: headShas[index],
-    })),
-  };
-  return createHash("sha256").update(canonicalJson(payload)).digest("hex");
 }
 
 /** Capture the exact clean workspace state after an applicable check suite passes. */
@@ -184,6 +159,9 @@ async function inspectWorkspaceForGateStep(
   "use step";
   const { Sandbox } = await import("@vercel/sandbox");
   const { getSandboxCredentials } = await import("../sandbox/credentials.js");
+  const { fingerprintWorkspaceState } = await import(
+    "./workspace-gate-fingerprint.js"
+  );
   const sandbox = await Sandbox.get({ sandboxId, ...getSandboxCredentials() });
 
   const manifestResult = await sandbox.runCommand("cat", [WORKSPACE_MANIFEST_PATH]);
@@ -247,20 +225,4 @@ function hasApplicableChecks(
       configured.has(`${repo.provider}:${repo.repoPath}`) &&
       (!repo.preAgentSha || repo.preAgentSha !== repo.headSha),
   );
-}
-
-function canonicalJson(value: unknown): string {
-  if (value === undefined) return "null";
-  if (value === null || typeof value !== "object") {
-    return JSON.stringify(value) ?? "null";
-  }
-  if (Array.isArray(value)) {
-    return `[${value.map((item) => canonicalJson(item)).join(",")}]`;
-  }
-  const entries = Object.entries(value as Record<string, unknown>)
-    .filter(([, item]) => item !== undefined)
-    .sort(([left], [right]) => left.localeCompare(right));
-  return `{${entries
-    .map(([key, item]) => `${JSON.stringify(key)}:${canonicalJson(item)}`)
-    .join(",")}}`;
 }
